requirements.txt
pydantic~=1.10.9
fastapi~=0.97.0
uvicorn~=0.22.0
python-dotenv~=1.0.0
psycopg2-binary~=2.9.6
peewee~=3.16.2